Add day 1 solution
My first attempt at part 2 did something like this:

    text = line.replace('zero', '0')
    text = text.replace('one', '1')
    text = text.replace('two', '2')
    ...

but one of the example inputs was "eightwothree". If you replace the digits in
this manner you convert that text to "eigh2three" then "eigh23" and pick 23 as
the calibration value for the line. But we really need to find the first written
out digit which is eight. This was solved with regex patterns matching:

    ^.*?(\d|zero|one|two|three|...|nine)'

which matches the shortest possible set of characters followed by a digit. The
digit matched is in group1 of the match.