Day08 cleaned up a bit more.
Part1 no longer raises a KeyError when run on example3.

Part2 makes a check to ensure that the cycle length from AAA to ZZZ
is the same as the second cycle from ZZZ to ZZZ. It is for all of cycles.
That means we can reliably use math.lcm() to figure out the total number
of steps to get all of the inputs to a **Z node.